Try running C2 Sync more in parallel
It's been timing out with over 9 hours for some sync jobs, which is
crazy. There is a bunch of inefficiencies in the dea-sync tool, but
perhaps we can get away with it a while longer with this.

It also appears to be opening a new database connection for every file,
and is also use GDAL to open every dataset while checking for validity.
YAML parsing is not insignificant either.