Collection 2 Sync: make more robust by limiting concurrent pbs jobs
We've been seeing regular timeouts of PBS Sync jobs for collection 2. We
haven't been able to pin down the root cause, but this is an attempt to
work around it by limiting the number of sync tasks running at once. In
case the limit is the Database server, or our connection to it.