Extend wait time for NCI Dataset Sync
They're *still* consistently running out of walltime when set to
9hr, but succeed when run again later. Bumping up to 12hr to try
to make them more reliable without digging in too deep to the
dea-sync code (which I know is doing *way* too much work).

We aren't using it elsewhere, and Collection 2 isn't going to be
around past the end of the year.