Try again to fix dataset sync
We were running in the copyq queue, but it only gives a single CPU
but does give higher performance access to the database. It is also
limited to 10 hour runtime, and we were hitting that or going too close.

This is an attempt to switch back to the normal queue, run with 4 cpus
and up to 10 hours, and hope for the best, that we're not overloading
the network restriction.