Rewrite Delete Dataset Utility DAG
When it was deleting many datasets, it was failing with:

    [2021-11-18 10:14:26,145] {pod_launcher.py:149} INFO - bash: line 10: /usr/bin/psql: Argument list too long

I've rewritten it to load the list of datasets straight into `psql`
rather than passing them as command line arguments.

This also now runs in a database transaction, so that either all the
records are deleted, or none of them are deleted.

The old method could leave the database in an inconsistent state.