Generate deterministic ids when uploading S2 ARD (#573)
Previously, this was a randomly generated ID, which can cause the same
dataset to be indexed as multiple datasets on AWS if the script is run
more than once.

The Landsat Collection 3 uploader uses the exact same ID on AWS, even
though the NBAR data is stripped out on upload. We could do the same
here, but I'm playing it safe and using different but stable ones.