Fix the Sentinel 2 Bulk Indexing DAG
It was failing due to trying to read missing Dataset documents,
thanks to a short circuit option of `--skip-check`, which made it
assume that dataset documents existed, due to the format of the
URL glob we passed.

I think we should be checking for failed uploads elsewhere, and
allow this DAG to only try to index Dataset documents which actually exist.

The error message it returns doesn't indicate that the dataset has only
been partially uploaded.